refactoring. moved utility functions to separate file to allow usage in other components in the future
